+++ core/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Added
+
+- `Store.findMembers` returns instances of classes and subclasses
+- `Store.observeFindMembers` returns an Observable and pushes new values when graph is updated
+- `Store.additions$` and `Store.removals$` are `rxjs` `Subject`s for listening to changes to the store
+- `observableGraph`  creates a new `rdflib.js` graph where changes to statements are observable through `rxjs` `Subject`s `additions$` and `removals$`
+
 ## 0.18.0
 
 ### ⚠ BREAKING CHANGES